# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Luna DevOps API
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/projects/{projectId}/applications/{applicationId}/deployment-targets:
    get:
      tags: [Deployments]
      summary: List deployment targets for an application
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApplicationId"
      responses:
        "200":
          description: Deployment target list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DeploymentTarget" }
    post:
      tags: [Deployments]
      summary: Create a deployment target
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeploymentTargetInput" }
      responses:
        "201":
          description: Created deployment target.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeploymentTarget" }
  /api/v1/projects/{projectId}/applications/{applicationId}/deployment-targets/{targetId}:
    put:
      tags: [Deployments]
      summary: Update a deployment target
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/TargetId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeploymentTargetInput" }
      responses:
        "200":
          description: Updated deployment target.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeploymentTarget" }
    delete:
      tags: [Deployments]
      summary: Delete a deployment target
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/TargetId"
      responses:
        "204":
          description: Deployment target deletion accepted and queued for asynchronous runtime cleanup.
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: lyd_session
      description: HttpOnly browser session cookie. Sensitive interactive operations do not accept personal access tokens as a substitute.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: Luna personal access token
  parameters:
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string }
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
    TargetId:
      name: targetId
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    DeploymentTargetInput:
      type: object
      description: Deployment target create/update payload. Other build and Kubernetes fields are accepted by the implemented endpoint; `webConsoleEnabled` is documented here because it has inherited policy semantics.
      additionalProperties: true
      properties:
        stage:
          type: string
          minLength: 2
          maxLength: 12
          pattern: "^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$"
          description: Immutable deployment stage identifier, unique within the application.
        buildDefinitionMode:
          type: string
          enum: [repository_dockerfile, template]
          default: repository_dockerfile
          description: Selects the repository Dockerfile or a platform-rendered template Dockerfile.
        buildTemplateId:
          type: string
          description: Required when buildDefinitionMode is template.
        buildTemplateVersion:
          type: string
          description: Immutable built-in template version. An empty value selects the current version.
        buildTemplateValues:
          type: string
          description: JSON object containing validated template parameters.
        buildVariables:
          type: object
          additionalProperties: { type: string }
          description: Optional deployment-level values that override matching application, project, and global keys.
        buildSecrets:
          type: object
          additionalProperties: { type: string }
          writeOnly: true
          description: Optional deployment-level secret updates. Existing keys with an empty value are retained; omitted keys are removed. Values are encrypted and never returned.
        webConsoleEnabled:
          type: [boolean, "null"]
          default: null
          description: "`null` inherits the project-space master switch and `false` disables Web Console for this deployment target. `true` is normalized to inheritance for compatibility and cannot bypass a disabled project-space switch."
    DeploymentTarget:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        projectId: { type: string }
        applicationId: { type: string }
        stage: { type: string, minLength: 2, maxLength: 12 }
        kubernetesName: { type: string }
        buildDefinitionMode: { type: string, enum: [repository_dockerfile, template] }
        buildTemplateId: { type: string }
        buildTemplateVersion: { type: string }
        buildTemplateValues: { type: string }
        webConsoleEnabled:
          type: [boolean, "null"]
          description: Nullable target restriction. Effective policy is `project.webConsoleEnabled && webConsoleEnabled != false`; a true value is treated as inheritance.
      required: [id, projectId, applicationId, stage, kubernetesName, webConsoleEnabled]
...